adjust config values to double

+++ trunk-recorder/global_structs.h
@@ -31,7 +31,7 @@ struct Config {
   bool new_call_from_update;
   bool debug_recorder;
   int debug_recorder_port;
-  int call_timeout;
+  double call_timeout;
   bool console_log;
   bool log_file;
   int control_message_warn_rate;
